Added Module sections to repo
Building each of the three windows as different modules before combining them into one program, the test module for the false loading window, and the image button have been made. NOTE: the start screen, and the game screen modules are the other two of the three, and have not been made yet. The button module is a sub-module of the game screen module.